part 1 of fix (kind of) for #2 form pages
The forms have been fixed to look a lot prettier. I had to hide the help text on the login page in order to do this (for some reason the fix displays the help text for the password help-text really strangely) but the error messages are descriptive and still work so I don't think thats a huge issue. So far I have fixed the editprof and the login pages. Going to work on the login page/hiding the password text next.