#14 Can send a simple email (also some other random stuff idk why its here)
Started playing with the emailing system. Made a temporary view for sending emails and if you go to  /emailpage it will send an email to my email address (you can change it to your email if you wanna try it out). We will need to change the sever name but when I tried to create a new server that was called DukeConversations it keeps declining my signup as being "high risk" so for now my Kimberly3 server will have to do.

Am struggling with git. Keep getting errors when I try to pull origin.